+++ Holloman_DesignPatternHomework/Scoop.cs
@@ -9,7 +9,7 @@ namespace Holloman_DesignPatternHomework
     public class Scoop : IngerdientDecorator
     {
         public string OrderName { get; set; }
-        public Scoop(double cost, string name, IngerdientDecorator obj, string order) : base(cost, name, null)
+        public Scoop(decimal cost, string name, IngerdientDecorator obj, string order) : base(cost, name, null)
         {
             OrderName = order;
         }
